rename chain to usecase
