+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/education/ProfileEducationCrud.kt
@@ -3,10 +3,10 @@ package com.lpfun.backend.profile.domain.education
 import com.lpfun.backend.common.model.profile.education.ProfileEducationContext
 
 class ProfileEducationCrud {
-    private val getChain = ProfileEducationGetChain()
-    private val createChain = ProfileEducationCreateChain()
-    private val updateChain = ProfileEducationUpdateChain()
-    private val deleteChain = ProfileEducationDeleteChain()
+    private val getChain = ProfileEducationGetUseCase()
+    private val createChain = ProfileEducationCreateUseCase()
+    private val updateChain = ProfileEducationUpdateUseCase()
+    private val deleteChain = ProfileEducationDeleteUseCase()
 
     suspend fun get(context: ProfileEducationContext) = getChain.execute(context)
     suspend fun create(context: ProfileEducationContext) = createChain.execute(context)
