+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/education/ProfileEducationDeleteUseCase.kt
@@ -1,12 +1,13 @@
 package com.lpfun.backend.profile.domain.education
 
 import com.lpfun.backend.common.model.profile.base.ProfileContextStatus
-import com.lpfun.backend.common.model.profile.base.stub.ProfileStubUpdate
+import com.lpfun.backend.common.model.profile.base.stub.ProfileStubDelete
+import com.lpfun.backend.common.model.profile.education.ProfileEducation
 import com.lpfun.backend.common.model.profile.education.ProfileEducationContext
 import com.lpfun.backend.profile.domain.cor.IExec
 import com.lpfun.backend.profile.domain.cor.cor
 
-class ProfileEducationUpdateChain : IExec<ProfileEducationContext> {
+class ProfileEducationDeleteUseCase : IExec<ProfileEducationContext> {
     override suspend fun execute(ctx: ProfileEducationContext) = chain.execute(ctx.apply {
 
     })
@@ -20,11 +21,11 @@ class ProfileEducationUpdateChain : IExec<ProfileEducationContext> {
 
             // Обработка стабов
             processor {
-                condition { stubCaseUpdate != ProfileStubUpdate.NONE }
+                condition { stubCaseDelete != ProfileStubDelete.NONE }
                 handler {
-                    condition { stubCaseUpdate == ProfileStubUpdate.RUNNING }
+                    condition { stubCaseDelete == ProfileStubDelete.RUNNING }
                     exec {
-                        responseProfile = requestProfile
+                        responseProfile = ProfileEducation()
                         responseProfileStatus = ProfileContextStatus.FINISHING
                     }
                 }
