+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/personal/ProfilePersonalCrud.kt
@@ -3,10 +3,10 @@ package com.lpfun.backend.profile.domain.personal
 import com.lpfun.backend.common.model.profile.personal.ProfilePersonalContext
 
 class ProfilePersonalCrud {
-    private val getChain = ProfilePersonalGetChain()
-    private val createChain = ProfilePersonalCreateChain()
-    private val updateChain = ProfilePersonalUpdateChain()
-    private val deleteChain = ProfilePersonalDeleteChain()
+    private val getChain = ProfilePersonalGetUseCase()
+    private val createChain = ProfilePersonalCreateUseCase()
+    private val updateChain = ProfilePersonalUpdateUseCase()
+    private val deleteChain = ProfilePersonalDeleteUseCase()
 
     suspend fun get(context: ProfilePersonalContext) = getChain.execute(context)
     suspend fun create(context: ProfilePersonalContext) = createChain.execute(context)
