+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/personal/ProfilePersonalUpdateUseCase.kt
@@ -1,12 +1,12 @@
 package com.lpfun.backend.profile.domain.personal
 
 import com.lpfun.backend.common.model.profile.base.ProfileContextStatus
-import com.lpfun.backend.common.model.profile.base.stub.ProfileStubCreate
+import com.lpfun.backend.common.model.profile.base.stub.ProfileStubUpdate
 import com.lpfun.backend.common.model.profile.personal.ProfilePersonalContext
 import com.lpfun.backend.profile.domain.cor.IExec
 import com.lpfun.backend.profile.domain.cor.cor
 
-class ProfilePersonalCreateChain : IExec<ProfilePersonalContext> {
+class ProfilePersonalUpdateUseCase : IExec<ProfilePersonalContext> {
     override suspend fun execute(ctx: ProfilePersonalContext) = chain.execute(ctx.apply {
 
     })
@@ -20,13 +20,11 @@ class ProfilePersonalCreateChain : IExec<ProfilePersonalContext> {
 
             // Обработка стабов
             processor {
-                condition { stubCaseCreate != ProfileStubCreate.NONE }
+                condition { stubCaseUpdate != ProfileStubUpdate.NONE }
                 handler {
-                    condition { stubCaseCreate == ProfileStubCreate.RUNNING }
+                    condition { stubCaseUpdate == ProfileStubUpdate.RUNNING }
                     exec {
-                        responseProfile = requestProfile.apply {
-                            profileId = "test-id"
-                        }
+                        responseProfile = requestProfile
                         responseProfileStatus = ProfileContextStatus.FINISHING
                     }
                 }
